fix: Add merge reconciliation retry cap and worktree cleanup
Merging state had no max retry count, causing infinite retry loops when
merge conflicts couldn't be resolved. This adds 5 fixes:

1. Max retry cap (3 attempts) — escalates to MergeConflict after limit
2. Registry guard — skips agent re-spawn if one is already running
3. Auto-complete fallback — logs warning when task stays stuck
4. Timeout increase — 180s → 300s for conflict resolution
5. Worktree cleanup — aborts stale rebase/merge and removes symlinks
   before recovery re-spawn

Also fixes pre-existing test (stale default quota) and clippy warning.

+++ src-tauri/tests/session_linking_integration.rs
@@ -107,7 +107,7 @@ async fn create_session_with_plan_and_proposals(
             selected: false,
             created_at: chrono::Utc::now(),
             updated_at: chrono::Utc::now(),
-            sort_order: i as i32,
+            sort_order: i,
         };
         proposal_ids.push(proposal.id.to_string());
         state.task_proposal_repo.create(proposal).await.unwrap();
